drmgr: workaround for qemu virtio hotplug devices
QEMU wants to hotplug slots on a single bus. The Power rpaphp module is broken
and only plays nicely when slots are on their own bus. Workaround this limitation
by using the /sys/bus/pci/rescan interface to discover new devices. The device
/sys/devices/pci[busid]/[bus:slot:fn]/remove sysfs entry is used to remove the
device. A new "-V" option is introduced to identify that this workaround for
pci virtio devices is to be used.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>