drmgr: add_nodes only adds half the nodes returned
The ibm,update-nodes rtas call returns an operation list whose first 4 bytes
are composed of a 1-byte opcode and 3-bytes denoting the number of operands
in the list. For the add nodes case each operand is actually composed of a
phandle/drc-index pair. The current code increments the count while processing
the list for each value instead of each value pair resulting in only half the
returned nodes being added.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>