drmgr: test suspend/migrate ops for in kernel device tree update
By testing the hibernate and migrate sysfs attributes drmgr can determine if a
device tree update is required. If the update is done by the kernel reading
the attribute file will return 1. Otherwise, it should either return 0 or be
unreadable to indicate drmgr needs to update the device tree from userspace.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>