drmgr: Use sysfs migration store to initiate migration when possible
The sysfs attribute /sys/kernel/mobility/migration has been present for
sometime, but never actually used by drmgr. It can be used to initiate a
migration operation in place of the existing faux ibm,supsend-me rtas call.
Further, the kernel is capable of doing the device tree update.

We can test the "migration" attribute to determine the course of action
drmgr should take. Newer kernels will return 1 indicating the kernel
does the update after a migration. If the attribute is unreadable or
returns 0 drmgr will resort to initiating the migration via rtas, and
communicate the post mobility device tree update through the
/proc/ppc64/ofdt interface to the kernel.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>