drmgr: allow attaching debugger early in process life cycle
The drmgr command is a transient process with short lifespan. While some
commands can be invoked on the commandline there are others like
migration that only make sense when invoked by the HMC. In these cases
its not feasable to attach a debugger to the running process as it has
probably already completed/crashed before a user has a chance.

Use an environment variable DRMGR_PAUSE that when defined will pause the
drmgr command in the main() function prior to branching out to the specific
dynamic reconfig command requested. The pid along with a message that drmgr
is awaiting a debugger will be logged in both /var/log/messages and
/var/log/drmgr. Once the user has attached a debugger they will need to
set a break point at the sleep(5) expression that is part of the while loop
in main(). After setting any other desired breakpoints in the drmgr code
the pause loop can be broken out of by setting debug_break=1.

Signed-off-by: Tyrel Datwyler <tyreld@linux.vnet.ibm.com>